+++ lib/services/service_locator.dart
@@ -2,6 +2,8 @@ import 'package:simutil/services/android_device_service.dart';
 import 'package:simutil/services/command_exec.dart';
 import 'package:simutil/services/ios_device_service.dart';
 import 'package:simutil/services/isolate_runner.dart';
+import 'package:simutil/services/plugin_registry_service.dart';
+import 'package:simutil/services/plugin_runner_service.dart';
 import 'package:simutil/services/settings_service.dart';
 import 'package:simutil/services/wifi_discovery_service.dart';
 
@@ -21,7 +23,10 @@ class ServiceLocator {
   );
   late final IOSDeviceService simctlService = IOSDeviceService(commandExec);
   late final SettingsService settingsService = SettingsServiceImpl();
-  late final WifiDiscoveryService wifiDiscoveryService = MdnsWifiDiscoveryService();
+  late final WifiDiscoveryService wifiDiscoveryService =
+      MdnsWifiDiscoveryService();
+  late final PluginRegistryService pluginRegistry = PluginRegistryServiceImpl();
+  late final PluginRunnerService pluginRunner = const PluginRunnerServiceImpl();
 
   Future<void> init() async => isolateRunner.init();